refactor(core): rename ticket related serializers to use new way of doing business
ref: #982 #983
